+++ lib/structs/events/encryption.cpp
@@ -12,9 +12,9 @@ from_json(const nlohmann::json &obj, Encryption &encryption)
         encryption.algorithm = obj.at("algorithm").get<std::string>();
 
         if (obj.contains("rotation_period_ms"))
-                encryption.rotation_period_ms = obj.at("rotation_period_ms").get<int>();
+                encryption.rotation_period_ms = obj.at("rotation_period_ms").get<uint64_t>();
         if (obj.contains("rotation_period_msgs"))
-                encryption.rotation_period_msgs = obj.at("rotation_period_msgs").get<int>();
+                encryption.rotation_period_msgs = obj.at("rotation_period_msgs").get<uint64_t>();
 }
 
 void
